added getDmax code
I forgot to add this to the repo initially.
It extracts the maximal diameter of the cells at each frame. It also estimates Dmax at t=Tc. This can be used for calculating the absolute constriction rates (nm/min).
Use the code within sDaDa_script to extract it during the analysis. Use getDmaxAllDatasets to extract it post factum when the data is already analyzed.